more exception handling and more junits

+++ src/main/java/com/tcs/assignment/customer/service/CustomerService.java
@@ -2,6 +2,8 @@ package com.tcs.assignment.customer.service;
 
 import com.tcs.assignment.customer.data.CustomerRepository;
 import com.tcs.assignment.customer.entity.Customer;
+import com.tcs.assignment.customer.exception.CustomerNotFound;
+import com.tcs.assignment.customer.exception.ErrorCode;
 import com.tcs.assignment.generate.model.CustomerRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -23,6 +25,9 @@ public class CustomerService {
     public Customer getCustomer(UUID id){
         logger.info("get customer by id in service class - {}", id.toString());
         Optional<Customer> customerOptional = customerRepository.findById(id);
+        if(customerOptional.isEmpty()){
+            throw new CustomerNotFound(ErrorCode.CUSTOMER_NOT_FOUND);
+        }
         return customerOptional.get();
     }
 
@@ -35,18 +40,29 @@ public class CustomerService {
 
     public Customer updateCustomer(String id, CustomerRequest customerRequest){
         logger.info("update customer started in service class - {}", id.toString());
+        Optional<Customer> custOpt = customerRepository.findById(UUID.fromString(id));
+        if(custOpt.isEmpty()){
+            throw new CustomerNotFound(ErrorCode.CUSTOMER_NOT_FOUND);
+        }
         // convert model to entity and update
         return customerRepository.save(customer(UUID.fromString(id), customerRequest));
     }
 
     public void deleteCustomer(UUID id){
         logger.info("delete customer started in service class - {}", id.toString());
+        Optional<Customer> custOpt = customerRepository.findById(id);
+        if(custOpt.isEmpty()){
+            throw new CustomerNotFound(ErrorCode.CUSTOMER_NOT_FOUND);
+        }
         customerRepository.deleteById(id);
     }
 
     public Customer getCustomerByName(String name){
         logger.info("get customer by name {}", name);
         List<Customer> customers = customerRepository.findByName(name);
+        if(customers == null || customers.isEmpty()){
+            throw new CustomerNotFound(ErrorCode.CUSTOMER_NOT_FOUND);
+        }
         return customers.get(0);
     }
 
@@ -54,6 +70,9 @@ public class CustomerService {
     public Customer getCustomerByEmail(String email){
         logger.info("get customer by email {}", email);
         List<Customer> customers = customerRepository.findByEmail(email);
+        if(customers == null || customers.isEmpty()){
+            throw new CustomerNotFound(ErrorCode.CUSTOMER_NOT_FOUND);
+        }
         return customers.get(0);
     }
